+++ cborg/tests/Main.hs
@@ -7,6 +7,7 @@ import qualified Tests.UnitTests  as UnitTests
 import qualified Tests.Properties as Properties
 import qualified Tests.Boundary   as Boundary
 import qualified Tests.ByteOffset as ByteOffset
+import qualified Tests.ByteSpan   as ByteSpan
 import qualified Tests.Canonical  as Canonical
 import qualified Tests.Regress    as Regress
 import qualified Tests.UTF8       as UTF8
@@ -22,6 +23,7 @@ tests =
     , UnitTests.testTree
     , Properties.testTree
     , ByteOffset.testTree
+    , ByteSpan.testTree
     , Boundary.testTree
     , Canonical.testTree
     , Regress.testTree